Add a post mortem template
This commit adds a post mortem template.

While the initial requirement for this was simply to make post mortems
easier to write, the structure of the yaml is one that can be parsed
easily.

Further, the goal with such structuring is to parse the post mortems
over time. There is work from Allspaw et. al which indicates that being
able to examine the post mortems in aggregate will generate insights
from the system that prove useful for making higher level decisions and
assigning budget.

This commit does not go all the way to such structuring, but does
provide tooling that is useful in and of itself, with the bonus of semi
structuring the data.

As the system evolves more for the purpose of human consumption a better
data model for post mortems should become more apparent.