Introduce the notion of "help" to post mortems
Post mortems are a pedagogical tool for understanding how a given
failure occurred. However, they still imply a set of knowledge about the
domain of failure analysis that is perhaps not universal across a given
organisation.

This commit introduces the primitive of "help", which provides some
"intellectual guard rails" and references to help newer users understand
the construct of a post mortem, rather than just its literal reflection
following an incident.