Restructure the lifecycle
Currently there are several primitives that bound the lifecycle in time.
There's:

- Start
- Finish
- What happens in the middle
- Closure, kindof.

And so fourth. However, they're all primitives that have their roots in
some event over time.

This commit restructures the application such that the timeline is much
clearer. This should make parsing the logs looking for correlations
across time much simpler and easier.